+++ packages/tts-editor/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Adds command to add UI file from object view
 - Adds command to get objects and save and play to object view toolbar
+- Adds command to get and update the current state of an object
 - Adds walkthrough
 
 ## [0.3.0] - 2024-01-05